Fixed issue of the terrain's displacement strength not being reset when the user chooses to reset Sand Paper

+++ SandPaper/src/tabs/TerrainTab.java
@@ -51,7 +51,7 @@ public class TerrainTab
      * 
      * @return The default strength
      */
-    public double getDefaultStrength()
+    public int getDefaultStrength()
     {
         return DEFAULT_STRENGTH;
     }
